added delete with token verification

+++ Part4/bloglist/controllers/blogs.js
@@ -72,8 +72,19 @@ blogsRouter.put("/:id", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id)
-	response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" })
+	}
+
+	const blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() === decodedToken.id) {
+		await Blog.findByIdAndDelete(request.params.id)
+		response.status(204).end()
+	}else{
+		return response.status(401).json({ error: "Unauthorized to delete the blog" })
+	}
 })
 
 module.exports = blogsRouter
